# Fix iOS streaming display and UI state management issues