## 🎯 **Overview**
This PR fixes multiple iOS streaming display and UI state management issues that were preventing real-time content updates from appearing in the chat interface.

## 🐛 **Issues Fixed**

### **Real-time Streaming Display**
- ✅ Content now appears immediately when streaming starts instead of only after completion
- ✅ Reduced StreamController throttling from 8ms to 2ms for more responsive updates 
- ✅ Removed guard clause that was blocking empty content updates from being processed
- ✅ Fixed timing race condition between `handleCompletion()` and `handleStreamContent()`

### **Send Button & Messaging Flow**
- ✅ Send button now works immediately when clicked (removed `isStreaming` guard)
- ✅ Users can send concurrent messages while previous responses are still streaming
- ✅ Send button only disabled during network requests (`isLoading`), not during streaming display

### **Progress Indicator Issues** 
- ✅ Fixed inappropriate typing indicator showing on app startup
- ✅ Typing indicator only appears during actual loading states
- ✅ Eliminated complex nested UI logic that was preventing content display

## 📋 **Technical Changes**

### **ChatView.swift**
- Simplified streaming UI logic to match working stream-to-screen demo pattern
- Separated streaming content display from typing indicator logic
- Fixed send button disabled condition to only check `isLoading`
- Removed complex nested VStack that was blocking content display

### **StreamController.swift**
- Reduced throttling delay from 8ms to 2ms for more responsive streaming

### **ChatViewModel.swift**
- Removed `isStreaming` guard from `sendMessage()` to allow concurrent messaging
- Fixed streaming state management logic to handle timing race conditions
- Added explicit handling for both content-present and empty-finished cases
- Ensured streaming state allows final content display even after `handleCompletion()`

## 🔧 **Root Cause Analysis**

The main issue was a **timing race condition**:
1. `handleCompletion()` sets `isStreaming = false` 
2. `handleStreamContent()` is called with final content but `isStreaming` is already false
3. UI condition `isStreaming && !content.items.isEmpty` fails, preventing display

**Solution**: Updated streaming state logic to handle timing correctly and simplified UI display logic.

## 🔍 **Verification**
- ✅ Build passes successfully
- ✅ No regressions in existing functionality
- ✅ Follows patterns from working stream-to-screen demo
- ✅ API streaming works correctly (verified in logs)

## 🎉 **Expected Behavior**
1. **Real-time streaming** - Content appears and updates as it arrives from API
2. **Immediate send functionality** - Send button works without delay
3. **Clean app startup** - No inappropriate progress indicators 
4. **Concurrent messaging** - Can send new messages while previous ones stream
5. **Proper state transitions** - UI states transition smoothly

## 📊 **Test Logs Analysis**
Based on the latest test logs:
- ✅ API receives and processes chunks correctly
- ✅ StreamController builds content successfully (`items=1 finished=false`)
- ✅ All streaming pipeline components work
- ✅ Only UI display logic needed the final fixes

---
**This is a Droid-assisted PR** - Comprehensive fixes for iOS streaming implementation based on the working stream-to-screen demo patterns.

---
**Factory Session:** https://app.factory.ai/sessions/dOPBSXBqqNzegt6kyxmX (created by sanoziec@gmail.com)